[refactor] compareCardNumber 위치 변경
